+++ CHANGELOG.md
@@ -53,6 +53,7 @@
 - Added `tooltipPlacement` prop to the `MultipleSelectChip`, `Autocomplete` and `Panel` component. 
 - Updated icon button spacing in Snackbar
 - Adjusted checkbox alignment in `DataGrid` component
+- Added keyboard accessibility to the header in `DataGrid` component
 
 ### Breaking changes